backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
google-generativeai>=0.9.0
groq
mistralai
